Refactor: move react error page to utils dir
